Build system improvements (#20)
* Clean up whitespace in Makefile

Signed-off-by: Stephen Kitt <steve@sk2.org>

* Use $(easyloggingdir) as the include directory

Signed-off-by: Stephen Kitt <steve@sk2.org>

* Install to $(DESTDIR)

Common practice for installation targets is to install to $(DESTDIR),
if it's defined. This is used by most packaging systems to install for
packaging purposes (instead of the final destination directories).

Signed-off-by: Stephen Kitt <steve@sk2.org>

* Use CXX instead of CC

Since the build uses C++, use $(CXX) and $(CXXFLAGS) instead of $(CC)
and $(CFLAGS); the latter are used for the C compiler.

Signed-off-by: Stephen Kitt <steve@sk2.org>

* Allow the build flags to be overridden

Packaging systems use CXX, CXXFLAGS, LDFLAGS etc. to define standard
flags used for package builds. This patch changes the build to
preserve the values of those externally-defined flags, augmenting them
if necessary.

Signed-off-by: Stephen Kitt <steve@sk2.org>